Corrected data_conversion, data_structures and host_computer library.

+++ src/host_computer.c
@@ -9,5 +9,5 @@ void host_comms_init(void){
 
 // In this function sending the data to the host computer is performed
 void send_host_packet(host_data_packet_t* host_data_packet){
-	printf("%f,%f,%f/n/r",host_data_packet->roll, host_data_packet->pitch, host_data_packet->yaw_vel);
+	printf("%f,%f,%f/n/r", (double)host_data_packet->roll, (double)host_data_packet->pitch, (double)host_data_packet->yaw_vel);
 }
